requirements.txt
# Updated requirements.txt with new dependencies for SaaS platform

# Existing dependencies
google-generativeai>=0.7.0
gspread
oauth2client
schedule
pytrends
python-dotenv
pandas
tweepy
beautifulsoup4
fastapi
uvicorn
pydantic
email-validator
python-multipart

# New SaaS dependencies
sqlalchemy>=2.0.0
psycopg2-binary  # PostgreSQL adapter
alembic  # Database migrations
python-jose[cryptography]  # JWT tokens
passlib[bcrypt]  # Password hashing
bcrypt==4.0.1  # Pin to version compatible with passlib
python-multipart  # Form data parsing
stripe  # Payment processing
sendgrid  # Email service
celery[redis]  # Task queue
redis  # Cache and message broker
boto3  # AWS S3 for file storage
openai>=1.0.0  # ChatGPT fallback
apscheduler  # Scheduled tasks
requests  # HTTP requests for scraping
